requirements.txt
pandas==1.3.1
requests==2.26.0
python-dotenv==0.19.0
vonage==2.5.5
pycoingecko